Add Trackable concern to create activities
