+++ app/controllers/concerns/trackable.rb
@@ -1,0 +1,13 @@
+module Trackable
+  extend ActiveSupport::Concern
+
+  included do
+    after_create :create_activity
+  end
+
+  private
+
+  def create_activity
+    activities.create!(user: user)
+  end
+end
