+++ app/models/jam.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Jam < ApplicationRecord
+  include Trackable
+
   belongs_to :room
   belongs_to :user
   has_one_attached :file
@@ -10,13 +12,7 @@ class Jam < ApplicationRecord
   has_many :activities, as: :subject, dependent: :destroy
   has_many :notifications, as: :target, dependent: :destroy
 
-  after_create :create_activity
-
   private
-
-  def create_activity
-    activities.create!(user: user)
-  end
 
   def content_type_is_audio
     return unless file.attached?
